fix: TaskListItem で使用するアイコンを変更
