+++ src/components/common/icons/Alarm.tsx
@@ -1,0 +1,20 @@
+import type { VFC } from "react";
+
+type Props = {
+  size?: number;
+};
+
+const Alarm: VFC<Props> = ({ size }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size ?? "24px"}
+    height={size ?? "24px"}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+  >
+    <path d="M0 0h24v24H0V0z" fill="none" />
+    <path d="M12.5 8H11v6l4.75 2.85.75-1.23-4-2.37zm4.837-6.19l4.607 3.845-1.28 1.535-4.61-3.843zm-10.674 0l1.282 1.536L3.337 7.19l-1.28-1.536zM12 4c-4.97 0-9 4.03-9 9s4.03 9 9 9 9-4.03 9-9-4.03-9-9-9zm0 16c-3.86 0-7-3.14-7-7s3.14-7 7-7 7 3.14 7 7-3.14 7-7 7z" />
+  </svg>
+);
+
+export default Alarm;
