+++ src/components/common/icons/AlarmOn.tsx
@@ -1,0 +1,20 @@
+import type { VFC } from "react";
+
+type Props = {
+  size?: number;
+};
+
+const AlarmOn: VFC<Props> = ({ size }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size ?? "24px"}
+    height={size ?? "24px"}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+  >
+    <path d="M0 0h24v24H0V0z" fill="none" />
+    <path d="M10.54 14.53L8.41 12.4l-1.06 1.06 3.18 3.18 6-6-1.06-1.06zm6.797-12.72l4.607 3.845-1.28 1.535-4.61-3.843zm-10.674 0l1.282 1.536L3.337 7.19l-1.28-1.536zM12 4c-4.97 0-9 4.03-9 9s4.03 9 9 9 9-4.03 9-9-4.03-9-9-9zm0 16c-3.86 0-7-3.14-7-7s3.14-7 7-7 7 3.14 7 7-3.14 7-7 7z" />
+  </svg>
+);
+
+export default AlarmOn;
