--- src/components/common/icons/TimerIcon.tsx
@@ -1,29 +1,0 @@
-import type { VFC } from "react";
-
-type Props = {
-  size?: number;
-};
-
-const TimerIcon: VFC<Props> = ({ size }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    enableBackground="new 0 0 24 24"
-    width={size ?? "24px"}
-    height={size ?? "24px"}
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    <g>
-      <rect fill="none" height="24" width="24" />
-    </g>
-    <g>
-      <g>
-        <g>
-          <path d="M15,1H9v2h6V1z M11,14h2V8h-2V14z M19.03,7.39l1.42-1.42c-0.43-0.51-0.9-0.99-1.41-1.41l-1.42,1.42 C16.07,4.74,14.12,4,12,4c-4.97,0-9,4.03-9,9s4.02,9,9,9s9-4.03,9-9C21,10.88,20.26,8.93,19.03,7.39z M12,20c-3.87,0-7-3.13-7-7 s3.13-7,7-7s7,3.13,7,7S15.87,20,12,20z" />
-        </g>
-      </g>
-    </g>
-  </svg>
-);
-
-export default TimerIcon;
