+++ src/components/common/icons/index.ts
@@ -1,7 +1,8 @@
+export { default as AlarmIcon } from "./Alarm";
+export { default as AlarmOnIcon } from "./AlarmOn";
 export { default as CircleIcon } from "./CircleIcon";
 export { default as DarkModeIcon } from "./DarkModeIcon";
 export { default as LightModeIcon } from "./LightModeIcon";
 export { default as LogoutIcon } from "./LogoutIcon";
 export { default as MenuIcon } from "./MenuIcon";
 export { default as PlayCircleIcon } from "./PlayCircleIcon";
-export { default as TimerIcon } from "./TimerIcon";
